moved from controller to model
