+++ src/MVC/Model.java
@@ -1,18 +1,37 @@
 package src.MVC;
+
 import src.Data.*;
 
+import javax.swing.filechooser.FileSystemView;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
+
+import static java.nio.charset.StandardCharsets.ISO_8859_1;
 
 public class Model {
     private ArrayList<ITag> allTags = new ArrayList<>();
     private ILock lockType;
     private HashMap<LocalDate, IDay> posts;
     private final LocalDate currentDate;
+    private final Charset charsetLatin;
+
     public Model() {
         posts = new HashMap<>();
         currentDate = LocalDate.now();
+        this.charsetLatin = ISO_8859_1;
+        init();
+    }
+
+    private void init(){
+        // load posts
+        createAppDirectory();
+        createPostsDirectory();
     }
 
     public void makePost(String text, int grade, ArrayList<ITag> tags, ArrayList<IMood> moods, ArrayList<ECondition> EConditions){
@@ -42,5 +61,82 @@ public class Model {
 
     public HashMap<LocalDate, IDay> getPosts() {
         return posts;
+    }
+
+    //functionality for saving orders goes here
+    public void savePosts() {
+        for (IDay post : posts.values()){
+            // create file name
+            String filename = getPostsDirectoryPath() + File.separatorChar + "post_" + post.getDate() + ".txt";
+
+            try{
+                FileOutputStream outputStream = new FileOutputStream(filename);
+                OutputStreamWriter outputWriter = new OutputStreamWriter(outputStream, charsetLatin);
+                String line;
+
+                // date
+                line = post.getDate() + "\n";
+                outputWriter.write(line);
+
+                // text
+                line = post.getText() + "\n";
+                outputWriter.write(line);
+
+                // grade
+                line = post.getGrade() + "\n";
+                outputWriter.write(line);
+
+                // moods
+                ArrayList<IMood> moods = post.getActiveMoods();
+                for (IMood mood : moods) {
+                    line = mood.getMoodName() + "|" + mood.getMoodRating() + "\n";
+                    outputWriter.write(line);
+                }
+
+                // tags
+                ArrayList<ITag> tags = post.getTags();
+                for (ITag tag : tags){
+                    line = tag.getTagID() + "\n";
+                    outputWriter.write(line);
+                }
+
+                // conditions
+                // TODO getConditions not available, fix during meeting (scared of merge conflicts :/)
+
+                outputWriter.flush();
+                outputWriter.close();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+    }
+
+
+    //create directories
+    private void createAppDirectory(){
+        File directory = new File(getAppDirectoryPath());
+        if (!directory.exists()){
+            directory.mkdirs();
+            System.out.println("App directory created");
+        } else {
+            System.out.println("App directory already exists");
+        }
+    }
+    private void createPostsDirectory(){
+        File directory = new File(getPostsDirectoryPath());
+        if (!directory.exists()){
+            directory.mkdirs();
+            System.out.println("Posts directory created");
+        } else {
+            System.out.println("Posts directory already exists");
+        }
+    }
+
+    //create directory & file names
+    private String getAppDirectoryPath(){
+        return FileSystemView.getFileSystemView().getDefaultDirectory().getPath() + File.separatorChar + "TDA367_Journal";
+    }
+    private String getPostsDirectoryPath(){
+        return this.getAppDirectoryPath() + File.separatorChar + "Posts";
     }
 }
